Move abs_path_if_local to new package_specifier module.
